CizayaEcommerce/CizayaProject/CizayaWebParent/CizayaBackEnd/src/main/java/com/cizaya/admin/category/CategoryRepository.java
package com.cizaya.admin.category;

import com.cizaya.common.entity.Category;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.PagingAndSortingRepository;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
public interface CategoryRepository extends PagingAndSortingRepository<Category, Integer> {

    //truy van cac danh muc cap cha- cap cao nhat
    @Query("SELECT c FROM Category c WHERE c.parent.id is NULL")
    public List<Category> findRootCategories();
}
